+++ src/main/java/com/hansung/treeze/persistence/MindmapSpecifications.java
@@ -12,11 +12,11 @@ import com.hansung.treeze.model.Mindmap_;
 
 public class MindmapSpecifications {
 	
-	public static Specification<Mindmap> isClassId(final int classId){
+	public static Specification<Mindmap> isClassId(final Long classId){
 		return new Specification<Mindmap>() {
 			@Override
 			public Predicate toPredicate(Root<Mindmap> root, CriteriaQuery<?> query, CriteriaBuilder cb){
-				return cb.equal(root.<Integer>get(Mindmap_.classId), classId);
+				return cb.equal(root.<Long>get(Mindmap_.classId), classId);
 			}
 		};
 	}
